+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Renamed `ForkedLegacyHttpsServerRunner` to `ForkedHttpsServerRunner`.
 - Changed `ForkedHttpsServerRunner` to run any HTTPS server that has a constructor like `MyHttpsServer(int, SSLProperties)` in a separate process so that it can have a separate SSL context.
 - Updated `logback-classic` to `1.5.27`.
+- Renamed package `json.jackson` to `json.jackson2` to better reflect the Jackson version used.
 
 ---